# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API documentation for SriRejeki IoT Greenhouse Monitoring",
        "title": "SriRejeki Backend API Documentation",
        "contact": {
            "name": "API Support",
            "email": "mail@hilmy.dev"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/auth": {
            "get": {
                "description": "Get user session data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user session data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
        "/auth/request-reset-password": {
            "post": {
                "description": "Send request reset user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Send request reset user password",
                "parameters": [
                    {
                        "description": "Phone number to send request reset password",
                        "name": "authRequestResetPasswordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authRequestResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "patch": {
                "description": "Reset user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset user's password",
                "parameters": [
                    {
                        "description": "Required data to reset password",
                        "name": "authResetPasswordRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "description": "Sign in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "Required sign in data",
                        "name": "authSignInRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authSignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/signup": {
            "post": {
                "description": "Sign up an account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign up an account",
                "parameters": [
                    {
                        "description": "Required account data",
                        "name": "authSignUpRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authSignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify-request-reset-password": {
            "post": {
                "description": "Verify request reset password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify request reset password",
                "parameters": [
                    {
                        "description": "Required data to verify request reset password",
                        "name": "authVerificationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify-signup": {
            "post": {
                "description": "Verify signed up account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify signed up account",
                "parameters": [
                    {
                        "description": "Code to verify signed up account",
                        "name": "authVerificationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
...
        "/mcu/auth": {
            "post": {
                "description": "Authenticate microcontroller | INTERNAL USE ON BACKEND ONLY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Authenticate microcontroller | INTERNAL USE ON BACKEND ONLY",
                "parameters": [
                    {
                        "description": "Required data to authenticate microcontroller",
                        "name": "microcontrollerAuthRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.authRequestResetPasswordRequest": {
            "type": "object",
            "required": [
                "phonenumber"
            ],
            "properties": {
                "phonenumber": {
                    "type": "string"
                }
            }
        },
        "auth.authResetPasswordRequest": {
            "type": "object",
            "required": [
                "code",
                "phonenumber"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
        "auth.authResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
        "auth.authSignInRequest": {
            "type": "object",
            "required": [
                "password",
                "phonenumber"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
        "auth.authSignUpRequest": {
            "type": "object",
            "required": [
                "name",
                "password",
                "phonenumber"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
        "auth.authVerificationRequest": {
            "type": "object",
            "required": [
                "code",
                "phonenumber"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
...
        "mcu.microcontrollerAuthRequest": {
            "type": "object",
            "required": [
                "deviceId"
            ],
            "properties": {
                "deviceId": {
                    "type": "string"
                }
            }
        },
...
        "mcu.microcontrollerResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
...
    }
}